Junction bug fixes
This set appears to work for simply Y junction using  test case of 'y_channel_002'. A major change is the addition of a new output module that creates a "threaded" output file.  This prints data from each link on a line, with the faces threaded between the elements and the values at the upstream/downstream connections (including boundary conditions). This output is useful for debugging.